Bug fixes
- Fixed a bug where the Poisson's ratio was quoted incorrectly for some
iron materials in the system database

- Fixed a bug where the 'fieldFormatString' environment variable was not
applied to field data which was written during an analysis using
CONTINUE_FROM

- Fixed a bug in CONTINUE_FROM, where the accumulated fatigue life would
sometimes increase to infinity. This happened because the accumulated
fatigue life was calculated by summing the damage between jobs, but if
the damage value is small this would be rounded to zero depending on the
value of the 'fieldFormatString' environment variable. The accumulated
fatigue life is now calculated as the inverse of the sum of the inverse
of the fatigue life of each analysis job

- Updated toolboxes to reflect the latest documentation chapter numbers