Various fixes
- Fixed an issue in the proof stress estimation code where the incorrect
value would be calculated
- If the local material path is not set, Material Manager searches for
an entry containing \Data\material\local in PATHDEF if the location does
not already exist on the current path. If a match is found, this
location is automatically suggested as the default local path. This
fixes the issue where the default path would always appear to be
unavailable when launching Material Manager from the MATLAB App bar