Bug fix
Fixed a bug in the Uniaxial Strain-Life app where the fatigue limit
stress was calculated incorrectly. The value was originally calculated
as the product of the fatigue limit strain with the Young's Modulus.
This was considered appropriate since at the fatigue limit, the stress
and strain are roughly proportional. However, for loadings where the
stress is slightly above the fatigue limit, zero damage would be
reported when the life is finite. The fatigue limit stress is now
calculated using the Ramberg-Osgood nonlinear elastic relationship.