Bug fix
Fixed a bug where a strain-life analysis with CONTINUE_FROM would abort
with an error that the field data file from the previous analysis is
corrupt.

This was due to a programming error whereby the overlay algorithm would
only check for stress units for the WCA, WCM and WCDP fields.

The algorithm now compares the two jobs and aborts if a combination of
stress-based and strain-based algorithms is used. The code also performs
the necessary check to ensure that two strain-life analyses are
superimposed without the aforementioned error message.