Feature deprecation
The `nlMaterial` environment vairable has been deprecated. The user can
no longer specify the plasticity correction for stress-based fatigue
analysis techniques.

This decision was made because the implementation was very innacurate;
stresses were corrected for plasticity on a cycle-by-cycle basis. This
technique does not account for ratcheting, permanent hysteresis or
material memory. Furthermore, since the stress-life methodology is
linear elastic, it is unlikely that the user would require this type of
correction.

In cases where nonlinearity must be accounted for, the user should
specify a strain-based method.