Bug fix
- The Rosette Analysis app would sometimes crash on the first usage in
the current MATLAB session

- Some output files would remain open after a fatigue analysis, meaning
that they could not be deleted until MATLAB was closed. Any open files
are now closed at the end of the analysis (including after an error).
The issue may still persist if the analysis stops due to an exception
which was not caught by Quick Fatigue Tool (red error message on command
line)