Enhancement to FRF algorithm
- If the user specified "No damage in compression" in the material
properties and provided user-defined FRF data, the horizontal FRF
calculation would consider the compressive side of the envelope. This
could produce the incorrect value, or a value of -1.0 if the compressive
side of the envelope was undefined. The algorithm now uses the positive
side of the envelope to reflect the fact that the compressive envelope
does not controbute to damage, thus yielding a less conservative (and
more representative) FRFH result